Remove parsedContent from AgentMessageType and generate client side, if needed.
