+++ front/pages/api/poke/workspaces/[wId]/conversations/[cId]/render.ts
@@ -12,9 +12,9 @@ import { getJITServers } from "@app/lib/api/assistant/jit_actions";
 import { listAttachments } from "@app/lib/api/assistant/jit_utils";
 import { getSkillServers } from "@app/lib/api/assistant/skill_actions";
 import { withSessionAuthenticationForPoke } from "@app/lib/api/auth_wrappers";
-import { getSupportedModelConfig } from "@app/lib/api/models";
 import { Authenticator } from "@app/lib/auth";
 import type { SessionWithUser } from "@app/lib/iam/provider";
+import { getSupportedModelConfig } from "@app/lib/llms/model_configurations";
 import { SkillResource } from "@app/lib/resources/skill/skill_resource";
 import { generateRandomModelSId } from "@app/lib/resources/string_ids";
 import { tokenCountForTexts } from "@app/lib/tokenization";
@@ -203,7 +203,6 @@ async function handler(
         skipToolsValidation: false,
         actions: [],
         contents: [],
-        parsedContents: {},
         modelInteractionDurationMs: null,
         completionDurationMs: null,
         richMentions: [],
