+++ front/temporal/agent_loop/lib/types.ts
@@ -1,8 +1,8 @@
 import type { AgentActionSpecification } from "@app/lib/actions/types/agent";
-import type { AgentMessageContentParser } from "@app/lib/api/assistant/agent_message_content_parser";
 import type { AgentMessageEvents } from "@app/lib/api/assistant/streaming/types";
 import type { LLMErrorInfo } from "@app/lib/api/llm/types/errors";
 import type { Authenticator } from "@app/lib/auth";
+import type { AgentMessageContentParser } from "@app/lib/llms/agent_message_content_parser";
 import type { AgentMessageModel } from "@app/lib/models/agent/conversation";
 import type {
   AgentConfigurationType,
